requirements.txt
streamlit>=1.36
google-cloud-dialogflow>=2.24.0
google-auth>=2.35.0
protobuf>=4.25.3
